+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 ### 2015.07.05
 * Renamed `Releaf::TemplateFieldTypeMapper` to `Releaf::Core::TemplateFieldTypeMapper`
+* Renamed `Releaf::AssetsResolver` to `Releaf::Core::AssetsResolver`
 
 ### 2015.08.04
 * refactored `@searchable_fields`. Now you should override `#searchable_fields`