+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Changelog
-### 2015.07.05
+### 2015.08.05
 * Renamed `Releaf::TemplateFieldTypeMapper` to `Releaf::Core::TemplateFieldTypeMapper`
 * Renamed `Releaf::AssetsResolver` to `Releaf::Core::AssetsResolver`
 * Renamed `Releaf::ErrorFormatter` to `Releaf::Core::ErrorFormatter`